Move argument error out of main describe block
Every individual test will need to rely on a hiera.yaml and a
common.yaml file being present. We mock these expectations in a before
:each block. The ONE test that doesn't need them is the parse error that
fails early when a key argument isn't provided. This test has its own
describe block so the two configuration parameter expectation mocks
don't need to be replicated in EVERY test that is written.